Rewrite HDF5Dataset to use entirely custom code, remove borrowed code.
The version of HDF5Dataset I originally borrowed processed data in chunks. So, __getitem__ would not return a single image, but instead a chunk of images. This isn't really in line with how PyTorch recommends building custom datasets.

As well, the old code I borrowed made use of caching in a complicated, unclear way. This isn't appropriate for my use case, so removing it allows me to simplify the HDF5Dataset class significantly.