Modify HDF5Dataset and Train/Val split to use indexes.
Previously, a single dataset was created for Train AND Val sets, and the splitting was handled by the DataLoader class. This proved impractical if different transformations were desired for training/validation, so the val splitting was changed to be handled prior to Dataset creation, using indices into the h5 file.